test(worker): split worker test by concern
